refactor: migrated Ibatis example in dao-router to MyBatis
docs: documented example
