+++ dao-router/src/main/resources/mybatis/mybatis-config.xml
@@ -42,38 +42,33 @@
   =========================LICENSE_END==================================
   -->
 
-<!DOCTYPE sqlMapConfig
-PUBLIC "-//ibatis.apache.org//DTD SQL Map Config 2.0//EN"
-"http://ibatis.apache.org/dtd/sql-map-config-2.dtd">
-<!-- Always ensure to use the correct XML header as above! -->
-<sqlMapConfig>
+<!DOCTYPE configuration PUBLIC "-//mybatis.org//DTD Config 3.0//EN" "http://mybatis.org/dtd/mybatis-3-config.dtd">
+<configuration>
 
 	<!-- These settings control SqlMap configuration details, primarily to do with transaction
 		management. They are all optional (see the Developer Guide for more). -->
-	<settings
-		cacheModelsEnabled="false"
-		enhancementEnabled="true"
-		lazyLoadingEnabled="false"
-		maxRequests="512"
-		maxSessions="128"
-		maxTransactions="16"
-		useStatementNamespaces="false"
-	/>
+	<settings>
+		<setting name="cacheEnabled" value="false"/>
+	</settings>
 
 	<!-- Configure a datasource to use with this SQL Map using SimpleDataSource.
 		Notice the use of the properties from the above resource -->
-	<transactionManager type="JDBC">
-		<dataSource type="SIMPLE">
-			<property name="JDBC.Driver" value="org.hsqldb.jdbcDriver" />
-			<property name="JDBC.ConnectionURL" value="jdbc:hsqldb:hsql://localhost:9201/milyn-hsql-9201" />
-			<property name="JDBC.Username" value="sa" />
-			<property name="JDBC.Password" value="" />
-		</dataSource>
-	</transactionManager>
-
+	<environments default="smooks">
+		<environment id="smooks">
+			<transactionManager type="JDBC"/>
+			<dataSource type="UNPOOLED">
+				<property name="driver" value="org.hsqldb.jdbcDriver" />
+				<property name="url" value="jdbc:hsqldb:hsql://localhost:9201/milyn-hsql-9201" />
+				<property name="username" value="sa" />
+				<property name="password" value="" />
+			</dataSource>
+		</environment>
+	</environments>
 
 	<!-- Identify all SQL Map XML files to be loaded by this SQL map. Notice the paths
 		are relative to the classpath.-->
-	<sqlMap resource="ibatis/sql-map.xml" />
+	<mappers>
+		<mapper resource="mybatis/mapper.xml"/>
+	</mappers>
 
-</sqlMapConfig>
+</configuration>
